identity-service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
alembic
psycopg2-binary
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
pydantic-settings
pytest
pytest-asyncio
